+++ fillit.h
@@ -42,7 +42,7 @@ typedef struct		s_map
 int					valid(char *str, int size);
 t_block				*make_list(char *str, int size);
 t_block				*make_block(char *str, char letter);
-size_t				count_in_list(t_block *head);
+int				    count_in_list(t_block *head);
 void				free_list(t_block *block);
 t_map				*new_map(int map_size);
 void				print_map(t_map *map, int map_size);
@@ -50,11 +50,11 @@ void				free_map(t_map *map, int map_size);
 t_block				*get_cord(char const *str, t_block *block);
 void				shift_axis(t_block *block, int axis);
 t_block				*shift(t_block *block);
-size_t				get_map_size(size_t num);
+size_t				get_map_size(int num);
 int					overlap(t_map *map, t_block *block);
 void				place(t_block *block, t_map *map, char letter);
-int					inside(t_block *block, size_t map_size, char axis);
-int					solve_map(t_map *map, t_block *block, size_t map_size);
+int					inside(t_block *block, int map_size, char axis);
+int					solve_map(t_map *map, t_block *block, int map_size);
 void				finder(t_block *head);
 
 #endif
